Add files via upload
These wallpaper images contain a variety of important structures such as large scale edges, smooth areas and fine textures, etc. Particularly, the average size of the dataset is 1254×1881, which is much larger than those of Cadik dataset and CSDD dataset. The average sizes of Cadik dataset and CSDD dataset are 292×352 and 356×406, respectively. The large size makes the conversion to be much more difficult than the Cadik dataset and CSDD dataset. It contains greater color complexity.